requirements.txt
streamlit==1.33.0
openai==1.14.3
pyyaml==6.0.1
python-dotenv==1.0.1
altair==5.2.0